added

+++ graduationProject/frontend/src/component/basket-page/BasketPage.jsx
@@ -5,7 +5,6 @@ import { useNavigate } from "react-router-dom";
 const BasketPage = () => {
   const urlList = "http://localhost:8080/api/cart/get/1";
   const [cart, setCart] = useState(undefined);
-  const [sumMoneys, setSumMoneys] = useState(0);
 
   const [isCartUpdated, setIsCartUpdated] = useState(false);
   function toUpperCaseTR(str) {
@@ -23,20 +22,12 @@ const BasketPage = () => {
       .replace(/([iışğüçö])/g, (letter) => mapping[letter] || letter)
       .toUpperCase();
   }
-  let sumMoney = 0;
-  //  function deneme(){
-  //     cart.cartProducts.map((cartProduct) => {
-  //       sumMoney += cartProduct.product.salesPrice * cartProduct.salesQuantity;
-  //     });
-  //   }
 
   useEffect(() => {
-    console.log("USE EFFECT ÇALIŞTI");
     fetch(urlList)
       .then((data) => data.json())
       .then((cart) => {
         setCart(cart);
-
         console.log(cart);
       });
   }, [isCartUpdated]);
@@ -47,13 +38,56 @@ const BasketPage = () => {
   };
 
   const removeToCartProduct = (cartProduct) => {
-    let url = `http://localhost:8080/api/cart/remove/1/${cartProduct.product.productId}`;
+    if (cart.cartStatus != "COMPLETED") {
+      let url = `http://localhost:8080/api/cart/remove/1/${cartProduct.product.productId}`;
+      fetch(url, {
+        method: "DELETE",
+      }).then(() => {
+        setIsCartUpdated(!isCartUpdated);
+      });
+    } else alert("Sepetiniz onaylanmıştır!!");
+  };
+
+  const checkoutBasket = () => {
+    let url = `http://localhost:8080/api/cart/checkout/1`;
     fetch(url, {
-      method: "DELETE",
+      method: "POST",
     }).then(() => {
       setIsCartUpdated(!isCartUpdated);
     });
+    if (cart.cartStatus != "COMPLETED") {
+      alert("Sepetiniz onaylanmıştır..");
+    }
   };
+
+  const reCheckoutBasket = () => {
+    let url = `http://localhost:8080/api/cart/reCheckout/1`;
+    fetch(url, {
+      method: "POST",
+    }).then(() => {
+      setIsCartUpdated(!isCartUpdated);
+    });
+    alert("Sepetiniz iptal edildi.");
+  };
+
+  const plusQuantity = (cartProduct) => {
+    if (cart.cartStatus != "COMPLETED") {
+      const url = `http://localhost:8080/api/cart/add/1/${cartProduct.product.productId}?quantity=1`;
+      fetch(url, { method: "POST" }).then(() => {
+        setIsCartUpdated(!isCartUpdated);
+      });
+    } else alert("Sepetiniz onaylanmıştır!!");
+  };
+
+  const minusQuantity = (cartProduct) => {
+    if (cart.cartStatus != "COMPLETED") {
+      const url = `http://localhost:8080/api/cart/minus/1/${cartProduct.product.productId}?quantity=1`;
+      fetch(url, { method: "POST" }).then(() => {
+        setIsCartUpdated(!isCartUpdated);
+      });
+    } else alert("Sepetiniz onaylanmıştır!!");
+  };
+
   return (
     <div>
       <div className="cart_section">
@@ -66,34 +100,45 @@ const BasketPage = () => {
                   <small className="text-muted fs-6 d-flex ms-2">
                     {" "}
                     (Sepetinizdeki ürün sayısı:{" "}
-                    {cart && cart.cartProducts.length} -- Sepet sahibi:{" "}
-                    {cart && toUpperCaseTR(cart.customerName)} ){" "}
+                    {cart &&
+                      cart.cartProducts
+                        .map((cartProduct) => cartProduct.salesQuantity)
+                        .reduce((a, b) => a + b, 0)}{" "}
+                    -- Sepet sahibi: {cart && toUpperCaseTR(cart.customerName)}{" "}
+                    ){" "}
                   </small>
                 </div>
                 {cart &&
-                  cart.cartProducts.map((cartProduct, index) => (
-                    <div className="cart_items" key={index}>
-                      <ul className="cart_list">
-                        <i
-                          onClick={() => removeToCartProduct(cartProduct)}
-                          style={{ cursor: "pointer" }}
-                          className="bi bi-trash-fill mt-2 me-2 fs-5 d-flex justify-content-end"
-                        ></i>
+                  cart.cartProducts
+                    .sort((a, b) =>
+                      a.product.productName.localeCompare(b.product.productName)
+                    )
+                    .map((cartProduct, index) => (
+                      <div className="cart_items" key={index}>
+                        <ul className="cart_list">
+                          <i
+                            onClick={() => removeToCartProduct(cartProduct)}
+                            style={{ cursor: "pointer" }}
+                            className="bi bi-trash-fill mt-2 me-2 fs-5 d-flex justify-content-end"
+                          ></i>
 
-                        <li className="cart_item clearfix">
-                          <div className="cart_item_image">
-                            <img src="https://i.imgur.com/qqBRWD5.jpg" alt="" />
-                          </div>
-                          <div className="cart_item_info d-flex flex-md-row flex-column justify-content-between">
-                            <div className="cart_item_name cart_info_col">
-                              <div className="cart_item_title">
-                                Marka / Model
-                              </div>
-                              <div className="cart_item_text">
-                                {cartProduct.product.productName}
-                              </div>
+                          <li className="cart_item clearfix">
+                            <div className="cart_item_image">
+                              <img
+                                src="https://i.imgur.com/qqBRWD5.jpg"
+                                alt=""
+                              />
                             </div>
-                            <div className="cart_item_color cart_info_col">
+                            <div className="cart_item_info d-flex flex-md-row flex-column justify-content-between">
+                              <div className="cart_item_name cart_info_col">
+                                <div className="cart_item_title">
+                                  Marka / Model
+                                </div>
+                                <div className="cart_item_text">
+                                  {cartProduct.product.productName}
+                                </div>
+                              </div>
+                              {/* <div className="cart_item_color cart_info_col">
                               <div className="cart_item_title">Renk</div>
                               <div className="cart_item_text">
                                 <span
@@ -101,37 +146,55 @@ const BasketPage = () => {
                                 ></span>
                                 Silver
                               </div>
-                            </div>
-                            <div className="cart_item_quantity cart_info_col">
-                              <div className="cart_item_title">Miktar</div>
-                              <div className="cart_item_text">
-                                {cartProduct.salesQuantity}
+                            </div> */}
+                              <div className="cart_item_quantity cart_info_col">
+                                <div className="cart_item_title"> Miktar</div>
+                                <div className="cart_item_text">
+                                  <i
+                                    onClick={() => minusQuantity(cartProduct)}
+                                    style={{ cursor: "pointer" }}
+                                    className="bi bi-dash-circle-fill me-1"
+                                  ></i>{" "}
+                                  {cartProduct.salesQuantity}{" "}
+                                  <i
+                                    onClick={() => plusQuantity(cartProduct)}
+                                    style={{ cursor: "pointer" }}
+                                    className="bi bi-plus-circle-fill ms-1"
+                                  ></i>
+                                </div>
+                              </div>
+                              <div className="cart_item_price cart_info_col">
+                                <div className="cart_item_title">Fiyat</div>
+                                <div className="cart_item_text">
+                                  {cartProduct.product.salesPrice}
+                                </div>
+                              </div>
+                              <div className="cart_item_total cart_info_col">
+                                <div className="cart_item_title">Toplam</div>
+                                <div className="cart_item_text">
+                                  ₺
+                                  {cartProduct.product.salesPrice *
+                                    cartProduct.salesQuantity}
+                                </div>
                               </div>
                             </div>
-                            <div className="cart_item_price cart_info_col">
-                              <div className="cart_item_title">Fiyat</div>
-                              <div className="cart_item_text">
-                                {cartProduct.product.salesPrice}
-                              </div>
-                            </div>
-                            <div className="cart_item_total cart_info_col">
-                              <div className="cart_item_title">Toplam</div>
-                              <div className="cart_item_text">
-                                ₺
-                                {cartProduct.product.salesPrice *
-                                  cartProduct.salesQuantity}
-                              </div>
-                            </div>
-                          </div>
-                        </li>
-                      </ul>
-                    </div>
-                  ))}
+                          </li>
+                        </ul>
+                      </div>
+                    ))}
                 <div className="order_total">
                   <div className="order_total_content text-md-right">
                     <div className="order_total_title">Toplam Fiyat:</div>
                     <div className="order_total_amount">
-                      ₺{cart && sumMoneys}
+                      ₺
+                      {cart &&
+                        cart.cartProducts
+                          .map(
+                            (cartProduct) =>
+                              cartProduct.salesQuantity *
+                              cartProduct.product.salesPrice
+                          )
+                          .reduce((a, b) => a + b, 0)}
                     </div>
                   </div>
                 </div>
@@ -144,9 +207,33 @@ const BasketPage = () => {
                   >
                     Ürünlere Dön
                   </button>{" "}
-                  <button type="button" className="button cart_button_checkout">
-                    Sepeti Onayla
-                  </button>{" "}
+                  {cart && cart.cartStatus != "COMPLETED" ? (
+                    <button
+                      onClick={checkoutBasket}
+                      type="button"
+                      className="button cart_button_checkout"
+                    >
+                      Sepeti Onayla
+                    </button>
+                  ) : (
+                    <button
+                      type="button"
+                      className="buttonEnd cart_button_checkoutEnd"
+                      disabled
+                    >
+                      Sepet Onaylandı
+                    </button>
+                  )}
+                  {cart && cart.cartStatus == "COMPLETED" && (
+                    <button
+                      onClick={reCheckoutBasket}
+                      type="button"
+                      style={{ backgroundColor: "red" }}
+                      className="ms-4 buttonEnd cart_button_checkout"
+                    >
+                      Sepeti İptal Et
+                    </button>
+                  )}
                 </div>
               </div>
             </div>
